fix: 커뮤니티 버튼 재사용가능하게 수정
